add resetting tracking account
