+++ hedera-node/src/main/java/com/hedera/services/ledger/TransferLogic.java
@@ -21,7 +21,6 @@ package com.hedera.services.ledger;
  */
 
 import com.google.protobuf.ByteString;
-import com.google.protobuf.InvalidProtocolBufferException;
 import com.hedera.services.context.SideEffectsTracker;
 import com.hedera.services.context.properties.GlobalDynamicProperties;
 import com.hedera.services.exceptions.InvalidTransactionException;
@@ -36,30 +35,24 @@ import com.hedera.services.store.tokens.TokenStore;
 import com.hedera.services.store.tokens.views.UniqTokenViewsManager;
 import com.hedera.services.txns.validation.OptionValidator;
 import com.hederahashgraph.api.proto.java.AccountID;
-import com.hederahashgraph.api.proto.java.Key;
-import com.hederahashgraph.api.proto.java.ResponseCodeEnum;
 import com.hederahashgraph.api.proto.java.TokenID;
 import org.apache.commons.lang3.tuple.Pair;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import static com.hedera.services.ledger.properties.AccountProperty.ALREADY_USED_AUTOMATIC_ASSOCIATIONS;
 import static com.hedera.services.ledger.properties.AccountProperty.BALANCE;
 import static com.hedera.services.ledger.properties.AccountProperty.NUM_NFTS_OWNED;
 import static com.hedera.services.ledger.properties.AccountProperty.TOKENS;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_KEY_ENCODING;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.OK;
 
 @Singleton
 public class TransferLogic {
 	private static final List<AccountProperty> TOKEN_TRANSFER_SIDE_EFFECTS =
 			List.of(TOKENS, NUM_NFTS_OWNED, ALREADY_USED_AUTOMATIC_ASSOCIATIONS);
-	private static final List<AccountProperty> CREATION_SIDE_EFFECTS = List.of();
 	private final TransactionalLedger<AccountID, AccountProperty, MerkleAccount> accountsLedger;
 	private final TransactionalLedger<NftId, NftProperty, MerkleUniqueToken> nftsLedger;
 	private final TransactionalLedger<Pair<AccountID, TokenID>, TokenRelProperty, MerkleTokenRelStatus> tokenRelsLedger;
@@ -67,8 +60,7 @@ public class TransferLogic {
 	private final TokenStore tokenStore;
 	private final MerkleAccountScopedCheck scopedCheck;
 	private final UniqTokenViewsManager tokenViewsManager;
-	private final TransferCreationsFactory transferCreations;
-	private static Map<ByteString, AccountID> aliasAccountsMap = new HashMap<>();
+	private final AutoAccountCreationsFactory transferCreations;
 
 	@Inject
 	public TransferLogic(final TransactionalLedger<AccountID, AccountProperty, MerkleAccount> accountsLedger,
@@ -79,7 +71,7 @@ public class TransferLogic {
 			final UniqTokenViewsManager tokenViewsManager,
 			final GlobalDynamicProperties dynamicProperties,
 			final OptionValidator validator,
-			final TransferCreationsFactory transferCreations) {
+			final AutoAccountCreationsFactory transferCreations) {
 		this.accountsLedger = accountsLedger;
 		this.nftsLedger = nftsLedger;
 		this.tokenRelsLedger = tokenRelsLedger;
@@ -108,7 +100,7 @@ public class TransferLogic {
 			}
 		}
 
-		validity = autoCreateForAliasTransfers(autoCreateAliases);
+		validity = transferCreations.autoCreateForAliasTransfers(autoCreateAliases, accountsLedger);
 
 		if (validity == OK) {
 			adjustHbarUnchecked(changes);
@@ -117,21 +109,6 @@ public class TransferLogic {
 			dropPendingAutoCreations();
 			throw new InvalidTransactionException(validity);
 		}
-	}
-
-	private ResponseCodeEnum autoCreateForAliasTransfers(final List<ByteString> autoCreateAliases) {
-		try {
-			List<Key> aliasKeys = new ArrayList<>();
-			for (ByteString alias : autoCreateAliases) {
-				final var key = Key.parseFrom(alias);
-				aliasKeys.add(key);
-				//Need to check if it is primitive key
-			}
-			transferCreations.createAutoAccounts(aliasKeys);
-		} catch (InvalidProtocolBufferException ex) {
-			return INVALID_KEY_ENCODING;
-		}
-		return OK;
 	}
 
 	private void adjustHbarUnchecked(final  List<BalanceChange> changes) {
@@ -147,6 +124,7 @@ public class TransferLogic {
 
 	private void dropPendingAutoCreations() {
 		accountsLedger.undoCreations();
+		transferCreations.clearTempCreations();
 		sideEffectsTracker.resetTrackedAutoCreatedAccount();
 	}
